Colab is Python 3.6
Most of the modules install fine under Python3.6 however wayback
requires Python3.7. This is a problem for Colab which is Python3.6. For
now we can install wayback from a fork where the 3.7 restriction is
relaxed. It tests just fine. If this PR gets merged, and it is released
to PyPI we can go back to installing wayback as normal.

https://github.com/edgi-govdata-archiving/wayback/pull/28